Cleanly emit compound types
Currently structs with padding and unions are not emitting correctly,
this won't be difficult to fix but it makes sense to do so in a seperate
commit, so I have commented out those tests for now and will fix them next.

This also fixes a bug with structs sorting their members incorrectly.